Finalize windows package build

+++ scripts/install_etcher.windows.bat
@@ -1,3 +1,4 @@
-rmdir /Q /S %2\etcher-cli
-tools\7z\7za.exe x %1 -o"%2\etcher-cli" -aoa
-copy scripts\etcher-cli.windows.bat %2\etcher-cli\
+@echo off
+if exist "%2\etcher-cli" rmdir /Q /S %2\etcher-cli
+%2\tools\7z\7za.exe x %1 -o"%2\etcher-cli" -aoa
+xcopy %2\scripts\etcher-cli.windows.bat %2\etcher-cli\
